Extract useAudioDeviceTooltipSelection
